Part Three - Checkpoints, Jump Limits and Fully Painted Levels oh boy!
This one had quite a few changes I made - added in a simple checkpoint system with a KillBox Script in order to ensure that in the 'platforming' areas, player does not fall indefinitely. Player will hit the DeathZone trigger, and respawn at the respective Checkpoint.

Ensured a Box Collider at the start of the game to ensure player does not fall off map intentionally.

Tried to edit the collision settings of the custom tiles I imported by editing the Custom Physics Shape in the Sprite Editor, but for whatever reason, my custom borders are not applying to the tiles - this is especially apparent with the grassy downhill portion of my level, going down player has no problem, trying to go back up causes the player to get stuck unless they jump. Will need to come back to this.

In my PlayerController script, I added a simple raycast system to make a ground check and ensure that the player does not jump more than once mid air, until they land again. This is currently not the best way of implementing this, will come back to it to clean it up or find a better system entirely.

Finished painting level, but will likely extend the first level even further - but will need to change the collisions first as stated above. I suspect this has something to do with my Composite Collider 2D or Tilemap Collider 2D applied to my Tilemap Game Object under Grid in the Hierarchy.

Proud of this so far - Unity has lots of potential for my previous Construct project to be fully realized!

-E