Part Five - Core Animations Complete, Dive Attack & Attack Limitations
Managed to complete animations for the Run, Jump, Fall, Idle and Attack States, as well as a coded in 'isFalling' check that works both as a pre-requisite for the fall animation to play and my Dive Attack to be possible.

I had planned on trying to get a sprite explosion to play behind the player as well as a particle effect to play around the player after completing their dive attack in this update, though that will need some more work as it bogged down my performance and did not play with the method I had used. The dive attack can be performed by holding down on the Down Arrow key while falling, to accelerate the player faster to the ground.

However, I did code in some more limitations for the player during the attack animation. They are the followng:
-Player cannot cast while in mid air
-Player cannot move while casting
-Player cannot jump while casting

These should help in the future to keep things balanced.

Next on the to do list will be to add projectiles to the player's cast, implement the shockwave to the dive attack, and to start some minor polish to move on to Enemies and such.

-E.M