Part 6 1/2 - Player Two Placeholder!
Whoops! Almost forgot to add in my second player into the scene - meant to do this earlier.

Nedferatu can be found next to Crows Hints (my NPC meant to give the player hints on narrative and controls). Temporarily he is just idling aimlessly, but will later be moved - to be either in scene as a bystander near the beginning if the player chooses Vickyula (and vice versa depending on player choice), or to be removed from scene altogether. Though I do like the idea of him/her waiting behind...

I'll be pasting the description of the previous part below so my full description isn't lost in the repository:

As the title suggests - this one is more of a minor update.

I mostly added in static Enemy sprites (Ranger and Skelly) to act as placeholders for the time being for later enemy locations, and where they will be posted for the player to deal with.

I've also added in my Sanguine Relic sprite and applied it's Idle animation and turned it into a prefab - though for some reason I can't just drag and drop it into my scene to make more than one animated item prefabs in the scene. Will need to take another look at this to make sure that I don't go about making several animated collectibles the wrong way and to keep things clean!

Lastly, I've made one change to the level - I added in some 'spikes' that will act as a static trap between two rangers. I also placed a Vampiric Rage near my multiple platform area that looks like a pyramid - though no animation was applied to that pickup.

Next steps will be;
-Adding enemy AI
-Straightening out Prefab Pickup Anims
-Making Pickups, well, pickupable
-Cleaning up world collision with player