Part Seven - Improved Camera, Animation Driven Mechanics, Functional Player Projectiles
Quite a lot was added to this update - I'll list below the following that I managed to get working both on a standard and improved level.

The Player camera is no longer a child of the player in the Unity Hierarchy, and has been given both X and Y limitations to follow through Private Serialized Fields that can be tweaked at will. Specifically to ensure that the camera does not exceed my level limitations, and that platforming is made semi easier for the player, so that they can now see where they will land.

A projectile will now spawn after each attack made by the player, driven by a animation driven script. On a singular frame of the 'Cast' state, roughly around halfway through the animation, the player projectile will spawn in. No gravity is associated with this projectile, it will travel directly horizontal to the player's cast position. In tandem with my previous isCasting check - this will stop the player from moving, and the attack spell cannot be cast while in midair. I also managed to flip the sprite based on the player's current position - the projectile sprite will correctly mirror the player's position based on their movement through the scene.

Lastly, I already briefly mentioned animation driven actions, but utilizing the state machine I've removed my 'Dive Attack' from the PlayerController.cs script and placed it into it's very own script tied to the falling animation. I do plan on changing the player's attack mechanic to be the same as this dive attack - tied to it's own animation - though that will be coming in the future.

To do:
-Still need to edit collisions of the tiles to ensure that the player does not seem like they're gliding above the ground!
-Clean up and separate mechanics into their own state machine scripts to be driven by their respective animation (Cast Attack)
-Solving Pickup prefab and turning them into items that can be collected.
-Adding in simple enemy AI pathing, though actual interaction with the player will come later. For now, the enemies will walk, and possible target the player for those that have ranged capabilities, but interaction (damage, collision damage etc) will be put off for now.
-Finally, I'd like to see if it's possible for me to add player controls and give animations to my second character - while still only making one controllable at a time. This may also come later however.

-E.M