Part Nine - Enemy Animations, Cleaned Player Projectile, More World Decorations!
More of another minor update this time around - this time focused on some polish and some decoration!

Added in some decorative candles and turned them into prefabs - they can now be placed around the level at will to be used as minor pieces of decoration!

Changed the Player's projectile to a trigger - this lets the projectile itself sail through level geometry and not be stopped in place awkwardly. This should also help to make things easier with my damage system and such later on.

Added in some decorative statues and a far background layer to add in the building in the pit area with the second powerup!

Gave the enemies that I plan to give basic AI in the next update some minor idle animations to go along with the second player and hints NPC at the beginning.

To Do:
-Still need to add in Enemy AI!
-Still need to make static traps functional - this should be in tandem with the health and health bar system.
-Health bar and health system!
-Second Player functionality.

-E.M