Part Ten - Simple Pickups Randomizer!
This is a simple update - mostly just to add in the pickups script I started work on the other day. I've placed five empty game objects in my scene, labelled as 'RandPickupSlots', which when game mode is activated five of my two pickup types will randomly spawn.

I managed to get some cool additional features added in with the monobehaviour script that I wrote up too - the user can set the rarity for the Vampiric Rage pickup, or the sanguine relic to change how often they'll spawn in the random pickup slots in scene. Thought that was pretty cool after finally getting it to work!

Though this was a smaller update, I'll be working on getting my turret enemy and patrolling enemy working shortly after this update. Following that, I'll still need to;

-Figure out my health system - and get functional health bars for both players and enemies!
-Still need to give static traps functionality - but health system should come first to avoid doing the same work twice.
-Second Player functionality!

-E.M