Part Eleven - Enemy AI and Simple Enemy Health!
This update took a while - still with some things to fix. There were quite a lot of changes that I made to this one!

Firstly, we have a simple patrolling enemy - the Skelly! He'll walk back and forth between edge markers/colliders, and act as a melee enemy. He can also be squished by jumping on his head!

Then, we have the Ranger! This turret enemy will not only patrol an area similar to the Skelly - but will stop and begin to shoot arrows at the Player when spotting them, after the Player comes within range of their detection sphere. This will continue until the player leaves that detection sphere. The player *should* be able to attack and destroy this Ranger with their spell attack - but for whatever reason - I could not get that figured out. The Ranger seems to just eat the Player's spell projectile and not take any damage.

Once I get that Ranger issue fixed, which is a current priority - the following will be on the list to be worked on:

- A proper health system for the player!
- Still need to get static traps to be functional!
- Second player functionality!
- Menu Functionality!

See you space cowboy!

-E.M