Part Twelve - Life Functionality, Health Functionality and Main Menu/Death Menu Functionality
Got quite a few things working in this update!

Both the Skelly making contact with the player physically and the Ranger's arrows making contact with the player will cause the player to lose 'lives' or 'hits' - once the player runs out, the death animation will play.

The game will start at the pause menu - assuming the player has started the game on that scene - and when the death animation plays, they will be brought to the 'death screen' or game over menu. If the player presses escape on this menu - they will be brought back to the main menu, and pressing escape again will restart the game from the beginning.

Sanguine relics also work as giving the player additional lives/hits - up to a total or maximum of nine. This is reflected through the debug log.

Still not sure why the ranger refuses to take damage? Will need to look into this more.

Future updates:

- Pause Menu will be next on the list!
- Now that health system is working, implementing health bar sprite will be next big step to work in tandem with 'lives'/'hits'
- Need to figure out why Ranger won't take damage from the player's spell attack and disappear.
- Static traps should be able to be implemented with the 'hit' system. Give the traps a bounce to the player? Or take away all health at once? This should also apply to the pitfall areas.

-E.M