Part Thirteen - Buttons, Pause Menus and UI Elements
Here I've managed to get another fairly big update done - this one focuses on functional menus, both the main and pause menus to be specific. I've yet to make changes to the 'Death' menu, though that will be in my next update I believe to tie it all together.

Changes made in this update:
- Starting from the main menu scene has three functional buttons; One to start the game, one to go to a settings menu (which will have sliders and other options in the next update) and one to exit the game, which will stop the game view from running.

- Pressing the P key while in game will pause the game and everything happening, and it will show three buttons - all of which are functional.

- There is a functional UI element when in game tied to my Player's Lives - picking up relics will also add to this count, and taking hits will reduce the amount of lives the player has. When depleted, the player's death animation will play.

From here I'll be working on adding music and sound to my game - and polishing as much as I can for A2. It is looking like I may have to cut some things from here, though I will continue to work on this project after the term ends to get it to a presentable stage.

-E.M