Part Fourteen - Sounds, Sliders and More!
Got another quite large update here - this one with sounds and more!

Here are the changes that I've made;
- The Main Menu, Main Game and Death Screen all have music that play when in those screens respectively in game view!

- The Main Menu now has a credits screen, and the Settings menu in the main menu has volume sliders to change all sound levels through a mixer!

- Virtually all actions from both the player and enemies in game will cause sound effects to play, this includes the player firing, jumping, dying, squishing the skelly, picking up collectibles, the ranger's arrows being shot, the ranger being taken down, etc!

- Added some text to the end screen if you die as the player to prevent a single blank screen from appearing!

There are some other minor changes as well but these are the core changes for this update. I unfortunately had to cut some things before I could get them implemented for this update - but I plan on adding them later on in my own time!

-E.M